requirements.txt
prefect>=2.0.0
azure-storage-blob>=12.0.0
pandas>=1.3.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 